[feature] 投稿詳細表示

+++ app/controllers/public/post_image_comments_controller.rb
@@ -2,9 +2,10 @@ class Public::PostImageCommentsController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    post_image = PostImage.find(params[:post_image_id])
-    @new_comment = post_image.post_image_comments.new(post_image_comment_params)
+    @post_image = PostImage.find(params[:post_image_id])
+    @new_comment = PostImageComment.new(post_image_comment_params)
     @new_comment.user_id = current_user.id
+    @new_comment.post_image_id = @post_image.id
     return unless @new_comment.save
 
     @new_comment = PostImageComment.new
@@ -12,7 +13,8 @@ class Public::PostImageCommentsController < ApplicationController
 
   def destroy
     post_image_comment = PostImageComment.find(params[:id])
-    post_image_comment.destroy
+    post_image_comment.destroy if post_image_comment.user == current_user
+    @post_image = PostImage.find(params[:post_image_id])
   end
 
   private
